Only add request to cache if the response is valid
This is achieved by having query() accept a second argument,
a validate function which is passed the response. The function's duty
will be to validate the response and throw an error if something is
invalid, so that query() can fail.

This will be most useful for queryNextStops(), so that we don't end up
caching invalid responses.