Make the Navitia code more robust to data changes
A recent dataset change broke the integration completely, since we were
assuming that the data was in the network "network:RTP".

This commit patches the Navitia code to rely less on navitia.io so
that it is more robust.

This also allows full backwards compatibility to be maintained, as only
new routes are added. Old routes are kept unaltered.