ratp/api: Add check for non-array or empty nextStops
Looks like the internal API is having some serious issues, since it
replies with an empty nextStops array randomly… during normal and even
peak hours. To prevent caching bad responses, this commit adds a basic
sanity check: if we are during operating hours and there are no trains,
fail the response check and don't cache the response. The client will
get back an appropriate 503.

I wouldn't be surprised if this turns out to be a hardware issue,
considering the API sometimes likes to give out dates and times in the
past or in the future (yay, let's wait at the station for ten years).